==> churchill_processing/1 gapfill met churchill.Rmd <==
---
title: "Prep for Ameriflux - Churchill"
author: "Kyle Arndt"
date: "2024-02-15"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = 'C:/Users/karndt.WHRC/Desktop/sites/churchill/ameriflux submission/') #use this to set your directory
```

Libraries
```{r}
rm(list = ls())
library(data.table)
library(ggplot2)
library(zoo)
Sys.setenv(TZ = "UTC")
```

Load Data
```{r}
c22 = fread(input = "./cleaned_Churchill_Met_Flux_Clean_Ameriflux_2022.csv")
c23 = fread(input = "./cleaned_Churchill_Met_Flux_Clean_Ameriflux_2023.csv")

df = rbind(c22,c23,fill=T)
df$ts = df$ts-6*3600 #adjust timezone from UTC to CST

df = subset(df,df$ts >= as.POSIXct(x = "2022-08-07 00:00"))
```


gap fill short gaps in the met data first
```{r}
gap = 3

#energy fluxes ###########################
df$SHF_6_37_1_1_1.c = na.spline(df$SHF_6_37_1_1_1.c,maxgap = gap)
df$SHF_6_37_2_1_1.c = na.spline(df$SHF_6_37_2_1_1.c,maxgap = gap)

#meteorology and weather #################################
df$TA_2_1_1_1_1     = na.spline(df$TA_2_1_1_1_1,maxgap = gap)
df$PA_4_2_1_1_1.c   = na.spline(df$PA_4_2_1_1_1.c,maxgap = gap)
df$RH_19_3_1_1_1.c  = na.spline(df$RH_19_3_1_1_1.c,maxgap = gap)

#Radiation #####################################
df$LWIN_6_14_1_1_1.c  = na.spline(df$LWIN_6_14_1_1_1.c,maxgap = gap)
df$LWOUT_6_15_1_1_1.c = na.spline(df$LWOUT_6_15_1_1_1.c,maxgap = gap)
df$SWIN_6_10_1_1_1.c = na.spline(df$SWIN_6_10_1_1_1.c,maxgap = gap)
df$SWOUT_6_11_1_1_1.c = na.spline(df$SWOUT_6_11_1_1_1.c,maxgap = gap)

df$RN_6_5_1_1_1.c = na.spline(df$RN_6_5_1_1_1.c,maxgap = gap)

df$PPFD_7_21_1_1_1.c = na.spline(df$PPFD_7_21_1_1_1.c,maxgap = gap)
df$PPFDR_7_23_1_1_1.c = na.spline(df$PPFDR_7_23_1_1_1.c,maxgap = gap)

#Soil factors ###################################
#all swc in %
df$SWC_12_36_1_1_1 = na.spline(df$SWC_12_36_1_1_1*100,maxgap = gap)
df$SWC_12_36_2_1_1 = na.spline(df$SWC_12_36_2_1_1*100,maxgap = gap)
df$SWC_12_36_2_2_1 = na.spline(df$SWC_12_36_2_2_1*100,maxgap = gap)

#all soil temps deg C
df$TS_2_38_1_1_1 = na.spline(df$TS_2_38_1_1_1,maxgap = gap)
df$TS_2_38_2_1_1 = na.spline(df$TS_2_38_2_1_1,maxgap = gap)
df$TS_2_38_2_2_1 = na.spline(df$TS_2_38_2_2_1,maxgap = gap)

df$TS_2_38_3_1_1.c  = na.spline(df$TS_2_38_3_1_1.c,maxgap = gap)
df$TS_2_38_3_2_1.c  = na.spline(df$TS_2_38_3_2_1.c,maxgap = gap)
df$TS_2_38_3_3_1.c  = na.spline(df$TS_2_38_3_3_1.c,maxgap = gap)
df$TS_2_38_3_4_1.c  = na.spline(df$TS_2_38_3_4_1.c,maxgap = gap)
df$TS_2_38_3_5_1.c  = na.spline(df$TS_2_38_3_5_1.c,maxgap = gap)
df$TS_2_38_3_6_1.c  = na.spline(df$TS_2_38_3_6_1.c,maxgap = gap)
df$TS_2_38_3_7_1.c  = na.spline(df$TS_2_38_3_7_1.c,maxgap = gap)
df$TS_2_38_3_8_1.c  = na.spline(df$TS_2_38_3_8_1.c,maxgap = gap)
df$TS_2_38_3_9_1.c  = na.spline(df$TS_2_38_3_9_1.c,maxgap = gap)
df$TS_2_38_3_10_1.c = na.spline(df$TS_2_38_3_10_1.c,maxgap = gap)
df$TS_2_38_3_11_1.c = na.spline(df$TS_2_38_3_11_1.c,maxgap = gap)
df$TS_2_38_3_12_1.c = na.spline(df$TS_2_38_3_12_1.c,maxgap = gap)

df$TS_2_38_4_1_1.c  = na.spline(df$TS_2_38_4_1_1.c,maxgap = gap)
df$TS_2_38_4_2_1.c  = na.spline(df$TS_2_38_4_2_1.c,maxgap = gap)
df$TS_2_38_4_3_1.c  = na.spline(df$TS_2_38_4_3_1.c,maxgap = gap)
df$TS_2_38_4_4_1.c  = na.spline(df$TS_2_38_4_4_1.c,maxgap = gap)
df$TS_2_38_4_5_1.c  = na.spline(df$TS_2_38_4_5_1.c,maxgap = gap)
df$TS_2_38_4_6_1.c  = na.spline(df$TS_2_38_4_6_1.c,maxgap = gap)
df$TS_2_38_4_7_1.c  = na.spline(df$TS_2_38_4_7_1.c,maxgap = gap)
df$TS_2_38_4_8_1.c  = na.spline(df$TS_2_38_4_8_1.c,maxgap = gap)
df$TS_2_38_4_9_1.c  = na.spline(df$TS_2_38_4_9_1.c,maxgap = gap)
df$TS_2_38_4_10_1.c = na.spline(df$TS_2_38_4_10_1.c,maxgap = gap)
df$TS_2_38_4_11_1.c = na.spline(df$TS_2_38_4_11_1.c,maxgap = gap)
df$TS_2_38_4_12_1.c = na.spline(df$TS_2_38_4_12_1.c,maxgap = gap)
```

```{r}
summary(df$PPFD_7_21_1_1_1.c) #check on the data coverage after filling small gaps

#save off a copy of the newly merged data
write.csv(x = df,file = "./churchill_2022_2023.csv",row.names = F)
```

Add ERA5 data
```{r}
#load in era5 data for churchill for continuous met data (found in Ameriflux folder in bucket)
era = fread('./era5_churchill.csv')

#rename to match up timestamps and merge datasets
df$date = df$ts

#merge together
all = merge(era,df,by = 'date',all = T)
all = all[!duplicated(all$date),]
```

We're now going to merge era5 data with site level data and create regression adjusted values for each variable
RH
```{r}
rh.mod = lm(formula = all$RH_19_3_1_1_1.c ~ all$rh)
summary(rh.mod)

all$rh.eramod = all$rh*rh.mod$coefficients[2] + rh.mod$coefficients[1]   

ggplot(data = all)+
  geom_point(aes(date,rh,col='ERA5'))+
  geom_point(aes(date,RH_19_3_1_1_1.c,col='Site'))

ggplot(data = all,aes(rh,RH_19_3_1_1_1.c))+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

#AirT
```{r}
ggplot(data = all)+
  geom_point(aes(date,airt,col='ERA5'))+
  geom_point(aes(date,all$TA_2_1_1_1_1,col='Site'))

airt.mod = lm(formula = all$TA_2_1_1_1_1 ~ all$airt)
summary(airt.mod)

all$airt.eramod = all$airt*airt.mod$coefficients[2] + airt.mod$coefficients[1]     

ggplot(data = all,aes(airt,TA_2_1_1_1_1))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

#SoilT
```{r}
ggplot(data = all)+
  geom_point(aes(date,st1,col='ERA5.1'))+
  geom_point(aes(date,st2,col='ERA5.2'))+
  geom_point(aes(date,all$TS_2_38_3_2_1.c,col='Site'))

soilt.mod = lm(formula = all$TS_2_38_3_2_1.c ~ all$st2)
summary(soilt.mod)
soilt.mod$coefficients
all$tsoil.eramod = all$st2*soilt.mod$coefficients[2] + soilt.mod$coefficients[1]           

ggplot(data = all)+
  geom_point(aes(date,st2,col='ERA5'))+
  geom_point(aes(date,TS_2_38_3_2_1.c,col='Site'))+
  geom_point(aes(date,tsoil.eramod,col='Lin. Reg.'))
```

#Rg
```{r}
ggplot(data = all)+
  geom_point(aes(date,rad,col='ERA5'))+
  geom_point(aes(date,SWIN_6_10_1_1_1.c,col='Site'))

#zoom in on a time period to make sure timestamps line up and solar noon makes sense
ggplot(data = all)+
  geom_point(aes(date,rad,col='ERA5'))+
  geom_point(aes(date,SWIN_6_10_1_1_1.c,col='Site'))+
  scale_x_datetime(limits = as.POSIXct(c('2023-07-01','2023-07-5')))

rad.mod = lm(formula = all$SWIN_6_10_1_1_1.c ~ all$rad)
summary(rad.mod)

all$rad.eramod = all$rad*rad.mod$coefficients[2] + rad.mod$coefficients[1]   

ggplot(data = all,aes(rad,SWIN_6_10_1_1_1.c))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

#ws
```{r}
ggplot(data = all)+
  geom_point(aes(date,ws,col='ERA5'))+
  geom_point(aes(date,WS_16_33_1_1_1,col='Site'))

ws.mod = lm(formula = all$WS_16_33_1_1_1 ~ all$ws)
summary(ws.mod)

all$ws.eramod = all$ws*ws.mod$coefficients[2] + ws.mod$coefficients[1]    

ggplot(data = all,aes(ws,WS_16_33_1_1_1))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

LE
```{r}
ggplot(data = all)+
  geom_point(aes(date,le,col='ERA5'))+
  geom_point(aes(date,LE.c,col='Site'))

le.mod = lm(formula = all$LE.c ~ all$le)
summary(le.mod)

all$le.eramod = all$le*le.mod$coefficients[2] + le.mod$coefficients[1]    

ggplot(data = all,aes(le,LE.c))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

H
```{r}
ggplot(data = all)+
  geom_point(aes(date,h,col='ERA5'))+
  geom_point(aes(date,H.c,col='Site'))

h.mod = lm(formula = all$H.c ~ all$h)
summary(h.mod)

all$h.eramod = all$h*h.mod$coefficients[2] + h.mod$coefficients[1]    

ggplot(data = all,aes(h,H.c))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

Pressure
```{r}
all$pres = all$pres/1000

ggplot(data = all)+
  geom_point(aes(date,pres,col='ERA5'))+
  geom_point(aes(date,all$PA_4_2_1_1_1.c,col='Site'))

pres.mod = lm(formula = all$PA_4_2_1_1_1.c ~ all$pres)
summary(pres.mod)

all$pres.eramod = all$pres*pres.mod$coefficients[2] + pres.mod$coefficients[1]    

ggplot(data = all,aes(pres,PA_4_2_1_1_1.c))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```


#reduce down to churchill size and add to final data set
```{r}
eras = all[,c('date','rh.eramod','airt.eramod','ws.eramod','rad.eramod','tsoil.eramod','le.eramod','h.eramod','pres.eramod')]

df = df[!duplicated(df$date),]
df = merge(df,eras,by = 'date',all.x = T)

write.csv(x = df,file = './churchill_2022_2023_era.csv',row.names = F,quote = F)
```
